separate main loop in its own modules to reuse it in repeat game
