+++ src/css/main.css
@@ -1,8 +1,8 @@
 @import url('https://fonts.googleapis.com/css?family=Exo:400,600&display=swap');
 
 body {
-  background-color: #333334;
-  color: white;
+  background-color: #383838;
+  color: #d6d6d6;
   font-family: 'Exo', sans-serif;
 }
 .container {
@@ -51,17 +51,17 @@ body {
 }
 
 .empty {
-  color: #9f9d9b;
+  color: #4b4a4a;
+  font-size: .6rem;
 }
 
 .ship {
   color: tomato;
-  font-size: 1.2rem;
+  font-size: 1.3rem;
 }
 
 .missed {
   color: #9f9d9b;
-  font-size: .6rem;
 }
 
 .hit {
